factored out ekf pendulum test
